Added initial delay in test

+++ ambry-prioritization/src/main/java/com/github/ambry/replica/prioritization/ReplicationPrioritizationManager.java
@@ -111,7 +111,7 @@ public class ReplicationPrioritizationManager implements Runnable {
 
     // We require this delay to ensure StoreManager is able to init all stores with correct state.
     // Otherwise, on server start all partitions store will be in BOOTSTRAP
-    int initialDelay = replicationConfig.prioritizationSchedulerInitialDelayMinutes;
+    long initialDelay = replicationConfig.prioritizationSchedulerInitialDelayMinutes;
     // Schedule periodic runs for prioritization run
     this.scheduler.scheduleAtFixedRate(this, initialDelay, scheduleIntervalMinutes, TimeUnit.MINUTES);
